rearrange modules/files for IOLocal integration
