+++ .github/workflows/ci.yml
@@ -68,7 +68,7 @@ jobs:
         run: sbt ++${{ matrix.scala }} test
 
       - name: Compress target directories
-        run: tar cf targets.tar modules/kernel/target modules/fs2/target modules/model/target target modules/rate-sampling/target modules/dynamic-sampling-config/target modules/model-testkit/target modules/exporter-stream/target modules/base/target modules/meta/target modules/base-laws/target modules/cats-effect-io/target modules/inject/target modules/filtering/target modules/log-exporter/target modules/tail-sampling/target modules/core/target modules/exporter-common/target modules/dynamic-sampling/target modules/testkit/target project/target
+        run: tar cf targets.tar modules/kernel/target modules/base-io/target modules/fs2/target modules/model/target target modules/inject-io/target modules/rate-sampling/target modules/dynamic-sampling-config/target modules/model-testkit/target modules/exporter-stream/target modules/base/target modules/meta/target modules/base-laws/target modules/inject/target modules/filtering/target modules/log-exporter/target modules/tail-sampling/target modules/core/target modules/exporter-common/target modules/dynamic-sampling/target modules/testkit/target project/target
 
       - name: Upload target directories
         uses: actions/upload-artifact@v2
