+++ modules/base-io/src/main/scala/io/janstenpickle/trace4cats/base/context/io/instances.scala
@@ -1,0 +1,3 @@
+package io.janstenpickle.trace4cats.base.context.io
+
+object instances extends IOLocalContextInstances
