+++ modules/base/src/main/scala/io/janstenpickle/trace4cats/base/optics/Getter.scala
@@ -8,5 +8,5 @@ package io.janstenpickle.trace4cats.base.optics
 
 object Getter {
   def apply[S, A](_get: S => A): Getter[S, A] = _get(_)
-  def id[A]: Getter[A, A] = identity[A]
+  implicit def id[A]: Getter[A, A] = identity[A]
 }
