+++ modules/inject-io/src/main/scala/io/janstenpickle/trace4cats/inject/io/package.scala
@@ -1,0 +1,5 @@
+package io.janstenpickle.trace4cats.inject
+
+import _root_.io.janstenpickle.trace4cats.base.context.io.IOLocalContextInstances
+
+package object io extends IOLocalTraceInstances with IOLocalContextInstances
